+++ R/add_stat_label.R
@@ -233,6 +233,10 @@ add_stat_label.tbl_summary <- function(x, location = c("row", "column"), label =
 #' @rdname add_stat_label
 add_stat_label.tbl_svysummary <- add_stat_label.tbl_summary
 
+#' @export
+#' @rdname add_stat_label
+add_stat_label.tbl_ard_summary <- add_stat_label.tbl_summary
+
 .add_stat_label_default_label_arg <- function(x, statistic) {
   statistic |>
     imap(