+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Added argument `tbl_ard_summary(overall)`. When `TRUE`, the ARD is parsed into primary ARD and the Overall ARD and we run `tbl_ard_summary() |> add_overall()`. (#1940)
 
+* Added `add_stat_label.tbl_ard_summary()` method. (#1969)
+
 # gtsummary 2.0.2
 
 Updates to address regressions in the v2.0.0 release: